# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fantasysports OpenApi Documentation",
        "description": "Fantasysports RESTful API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://146.59.15.26:8080/api/v1",
            "description": "Fantasysports API Server",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "https"
                }
            }
        }
    ],
    "paths": {
...
        "/contests/upcoming": {
            "get": {
                "tags": [
                    "Contests"
                ],
                "summary": "Get Contests Upcoming",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ContestResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/balance": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User Balance",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "balance": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": "100.23"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User Profile",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept"
                    },
                    {
                        "$ref": "#/components/parameters/Content-Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProfileResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "405": {
                        "$ref": "#/components/responses/405"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ContestResource": {
                "title": "ContestResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Test Contest"
                    },
                    "status": {
                        "type": "integer",
                        "enum": [
                            1,
                            2,
                            3,
                            4,
                            5
                        ]
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "user",
                            "template"
                        ]
                    },
                    "contestType": {
                        "type": "string",
                        "enum": [
                            "fifty-fifty",
                            "head-to-head",
                            "multiplier",
                            "wta",
                            "top-three",
                            "custom"
                        ]
                    },
                    "expectedPayout": {
                        "type": "number",
                        "format": "double",
                        "example": "34.17"
                    },
                    "isPrizeInPercents": {
                        "type": "integer"
                    },
                    "maxEntries": {
                        "type": "integer",
                        "example": "12"
                    },
                    "maxUsers": {
                        "type": "integer",
                        "example": "12"
                    },
                    "minUsers": {
                        "type": "integer",
                        "example": "1"
                    },
                    "leagueId": {
                        "type": "integer",
                        "example": "5"
                    },
                    "startDate": {
                        "type": "integer",
                        "example": "1650112441"
                    },
                    "endDate": {
                        "type": "integer",
                        "example": "1650122541"
                    },
                    "details": {
                        "type": "string"
                    },
                    "entryFee": {
                        "type": "number",
                        "format": "double",
                        "example": "11.87"
                    },
                    "salaryCap": {
                        "type": "integer",
                        "example": "7000"
                    },
                    "prizeBank": {
                        "type": "number",
                        "format": "double",
                        "example": "90.45"
                    },
                    "prizeBankType": {
                        "type": "integer",
                        "enum": [
                            1,
                            2,
                            3,
                            5
                        ]
                    },
                    "customPrizeBank": {
                        "type": "number",
                        "format": "double",
                        "example": "71.23"
                    },
                    "maxPrizeBank": {
                        "type": "number",
                        "format": "double",
                        "example": "300.99"
                    },
                    "suspended": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "numEntries": {
                        "type": "integer",
                        "example": "1"
                    },
                    "numUsers": {
                        "type": "integer",
                        "example": "1"
                    },
                    "entries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContestUserResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "ContestUserResource": {
                "title": "ContestUserResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string",
                        "example": "fantasysports"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "budget": {
                        "type": "integer",
                        "example": "1350"
                    },
                    "date": {
                        "type": "integer",
                        "example": "1650112441"
                    },
                    "isWinner": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "place": {
                        "type": "integer",
                        "example": "1"
                    },
                    "prize": {
                        "type": "number",
                        "format": "double",
                        "example": "140.56"
                    },
                    "score": {
                        "type": "number",
                        "format": "double",
                        "example": "23.12"
                    }
                },
                "type": "object"
            },
...
            "ProfileResource": {
                "title": "ProfileResource",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "fantasysports"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@fantasysports.com"
                    },
                    "balance": {
                        "type": "number",
                        "format": "double",
                        "example": "100.23"
                    },
                    "dob": {
                        "type": "string",
                        "example": "1993-05-23",
                        "nullable": true
                    },
                    "countryId": {
                        "type": "integer"
                    },
                    "favTeamId": {
                        "type": "integer"
                    },
                    "favPlayerId": {
                        "type": "integer"
                    },
                    "languageId": {
                        "type": "integer"
                    },
                    "receiveNewsletters": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "receiveNotifications": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "avatarId": {
                        "type": "integer"
                    },
                    "isEmailConfirmed": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "invitedByUser": {
                        "type": "integer"
                    },
                    "isSham": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2022-04-01 12:00:05"
                    },
                    "updatedAt": {
                        "type": "string",
                        "example": "2022-04-02 11:59:41"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Resource not found."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "405": {
                "description": "Method Not Allowed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The current method is not supported for this route."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Server Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "Accept": {
                "name": "Accept",
                "in": "header",
                "description": "Accept header",
                "schema": {
                    "type": "string",
                    "example": "application/vnd.api+json"
                }
            },
            "Content-Type": {
                "name": "Content-Type",
                "in": "header",
                "description": "Content-Type header",
                "schema": {
                    "type": "string",
                    "example": "application/vnd.api+json"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer {{token}}**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}